+++ src/main/java/com/clinicarodriguez/clinicarodriguez/repository/EspecialidadesRepository.java
@@ -3,6 +3,6 @@ package com.clinicarodriguez.clinicarodriguez.repository;
 import com.clinicarodriguez.clinicarodriguez.model.Especialidades;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EspecialidadesRepository extends JpaRepository<Especialidades, Long> {
+public interface EspecialidadesRepository extends JpaRepository<Especialidades, Integer> {
 }
