+++ src/main/java/com/clinicarodriguez/clinicarodriguez/repository/MedicosEspecialidadesRepository.java
@@ -6,18 +6,18 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface MedicosEspecialidadesRepository extends JpaRepository<MedicosEspecialidades, Long> {
+public interface MedicosEspecialidadesRepository extends JpaRepository<MedicosEspecialidades, Integer> {
     
     // Buscar todas las especialidades de un médico
     @Query("SELECT me FROM MedicosEspecialidades me WHERE me.medico.mediId = :medicoId")
-    List<MedicosEspecialidades> findByMedicoId(@Param("medicoId") Long medicoId);
+    List<MedicosEspecialidades> findByMedicoId(@Param("medicoId") Integer medicoId);
     
     // Buscar todos los médicos de una especialidad
     @Query("SELECT me FROM MedicosEspecialidades me WHERE me.especialidad.espeId = :especialidadId")
-    List<MedicosEspecialidades> findByEspecialidadId(@Param("especialidadId") Long especialidadId);
+    List<MedicosEspecialidades> findByEspecialidadId(@Param("especialidadId") Integer especialidadId);
     
     // Verificar si existe una relación entre médico y especialidad
     @Query("SELECT COUNT(me) > 0 FROM MedicosEspecialidades me WHERE me.medico.mediId = :medicoId AND me.especialidad.espeId = :especialidadId")
-    boolean existsByMedicoAndEspecialidad(@Param("medicoId") Long medicoId, @Param("especialidadId") Long especialidadId);
+    boolean existsByMedicoAndEspecialidad(@Param("medicoId") Integer medicoId, @Param("especialidadId") Integer especialidadId);
 }
